Move pdl, azuread and redis up to app-file
